feat: add new branch for input, move from form
