+++ src/components/index.js
@@ -2,6 +2,7 @@ export * from "./avatar";
 export * from "./button";
 export * from "./feed";
 export * from "./form";
+export * from "./input";
 export * from "./layout";
 export * from "./modal";
 export * from "./notification";
